% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_autocorrelationFunction.R
\name{calculate_autocorrelationFunction}
\alias{calculate_autocorrelationFunction}
\title{Autocorrelation Function}
\usage{
calculate_autocorrelationFunction(observed)
}
\arguments{
\item{observed}{A numeric vector of observed values.}
}
\value{
An autocorrelation object.
}
\description{
Computes the autocorrelation function for the observed time series.
}
