% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_kendallTau.R
\name{calculate_kendallTau}
\alias{calculate_kendallTau}
\title{Kendall's tau}
\usage{
calculate_kendallTau(observed, predicted)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{predicted}{A numeric vector of predicted values.}
}
\value{
A numeric value representing Kendall's tau.
}
\description{
Calculates Kendall's tau, a measure of rank correlation between predicted and observed values.
}
\examples{
observed <- c(10, 15, 12, 8, 20)
predicted <- c(12, 16, 10, 7, 18)
model_efficiency(observed, predicted)

}
