% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mef.R
\name{calculate_mef}
\alias{calculate_mef}
\title{Calculate MEF}
\usage{
calculate_mef(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed values.}

\item{predicted}{Numeric vector of predicted values.}
}
\value{
The calculated MEF.
}
\description{
This function calculates the model efficiency (MEF) between observed and predicted values.
}
\examples{
observed <- c(1, 2, 3)
predicted <- c(1.1, 2.2, 3.3)
calculate_mef(observed, predicted)
}
