% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Metrics}
\usage{
calculate_metrics(df, observed, predicted)
}
\arguments{
\item{df}{A data frame containing the observed and predicted values.}

\item{observed}{The name of the column containing the observed values in \code{df}.}

\item{predicted}{The name of the column containing the predicted values in \code{df}.}
}
\value{
A data frame containing calculated metrics and plots.
}
\description{
This function calculates various evaluation metrics for observed and predicted values.
}
\examples{
df <- data.frame(observed = c(1, 2, 3), predicted = c(1.1, 2.2, 3.3))
calculate_metrics(df, "observed", "predicted")


}
