% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_modelSkillScore.R
\name{calculate_modelSkillScore}
\alias{calculate_modelSkillScore}
\title{Model Skill Score}
\usage{
calculate_modelSkillScore(observed, predicted)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{predicted}{A numeric vector of predicted values.}
}
\value{
A numeric value representing the model skill score.
}
\description{
Measures the skill of a model in predicting observed values.
}
