% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nmae.R
\name{calculate_nmae}
\alias{calculate_nmae}
\title{Calculate NMAE}
\usage{
calculate_nmae(observed, predicted)
}
\arguments{
\item{observed}{Numeric vector of observed values.}

\item{predicted}{Numeric vector of predicted values.}
}
\value{
The calculated NMAE.
}
\description{
This function calculates the normalized mean absolute error (NMAE) between observed and predicted values.
}
\examples{
observed <- c(1, 2, 3)
predicted <- c(1.1, 2.2, 3.3)
calculate_nmae(observed, predicted)
}
