% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylorD.R, R/taylorDiagram.R
\name{taylorD}
\alias{taylorD}
\title{Taylor Diagram}
\usage{
taylorD(observed, predicted, by = NULL, by.class = NULL, panel = TRUE)

taylorD(observed, predicted, by = NULL, by.class = NULL, panel = TRUE)
}
\arguments{
\item{observed}{A vector of observed values.}

\item{predicted}{A vector or data frame containing predicted values from one or more models.}

\item{by}{A vector with the same length as 'observed', containing values of the variable to condition the plots to.}

\item{by.class}{Character, "numeric" or "factor". If "numeric", data will be split in quantiles.}

\item{panel}{Logical, TRUE to plot segregated values in different panels, and FALSE to differentiate by model and 'by' variable within a single panel.}
}
\value{
A Taylor Diagram plot.

A Taylor Diagram plot.
}
\description{
This function creates a Taylor Diagram to compare predictive capacity (RMSE + correlation) of one or more models, with the option to segregate by a variable.

This function creates a Taylor Diagram to compare predictive capacity (RMSE + correlation) of one or more models, with the option to segregate by a variable.
}
\examples{
observed <- c(2, 4, 5, 7)
model1 <- c(3, 5, 4, 8)
model2 <- c(2, 4.5, 5, 7.2)
predicted <- as.data.frame(cbind(model1, model2))
year <- c(2000, 2000, 2001, 2001)
taylor(observed, model1)

observed <- c(2, 4, 5, 7)
model1 <- c(3, 5, 4, 8)
model2 <- c(2, 4.5, 5, 7.2)
predicted <- as.data.frame(cbind(model1, model2))
year <- c(2000, 2000, 2001, 2001)
taylor(observed, model1)

}
