% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_inspection.R
\name{visual_inspection}
\alias{visual_inspection}
\title{Visual Inspection Function}
\usage{
visual_inspection(df, observed, predicted, year, rasterVariable)
}
\arguments{
\item{df}{A data frame containing the data for visualization.}

\item{observed}{The column name of the observed fish biomass values in the data frame.}

\item{predicted}{The column name of the predicted fish biomass values in the data frame.}

\item{year}{The column name of the year information in the data frame.}

\item{rasterVariable}{The column name of the raster variable information in the data frame.}
}
\value{
A grid of plots displaying comparisons between observed and predicted fish biomass mean values.
}
\description{
This function creates aesthetically pleasing visualizations for comparing observed and predicted fish biomass mean values over time and with a raster variable.
Plots include temporal comparisons, raster variable-wise comparisons, and a scatter plot of observed vs. predicted fish biomass means.
}
\examples{
# Creating an example data frame with more natural fish biomass data
set.seed(123)  # Set seed for reproducibility
years <- 2000:2022
num_rows <- length(years) * 10
df <- data.frame(
  year = rep(rep(years, each = 10), times = 3),
  biomass_observed = 100 - 0.5 * rep(years, each = 10) + rnorm(num_rows, sd = 10) + 1000,
  biomass_predicted = 95 - 0.5 * rep(years, each = 10) + rnorm(num_rows, sd = 5) + 1000,
  variable = sample(0:2000, num_rows, replace = TRUE)
)
# Calling the visual_inspection function
visual_inspection(df, "biomass_observed", "biomass_predicted", "year", "variable")

}
