% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indexOfAgreement.R
\name{calculate_indexOfAgreement}
\alias{calculate_indexOfAgreement}
\title{Index of Agreement}
\usage{
calculate_indexOfAgreement(observed, predicted)
}
\arguments{
\item{observed}{A numeric vector of observed values.}

\item{predicted}{A numeric vector of predicted values.}
}
\value{
A numeric value representing the index of agreement.
}
\description{
Quantifies the agreement between predicted and observed values.
}
